+++ lib/ua_inspector/supervisor.ex
@@ -17,11 +17,7 @@ defmodule UAInspector.Supervisor do
 
   @doc false
   def init(_default) do
-    :ok =
-      case Config.get(:init) do
-        nil -> :ok
-        {mod, fun} -> apply(mod, fun, [])
-      end
+    :ok = Config.init_env()
 
     options = [strategy: :one_for_one, name: __MODULE__]